refactored Covid EHP Disclaimer into own component
